# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Gestión de Recetas","description":"API para gestionar recetas de yogurt"}],"paths":{"/api/recipes/{id}":{"get":{"tags":["Gestión de Recetas"],"summary":"Obtener receta","description":"Obtiene una receta por ID","operationId":"getRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Receta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"200":{"description":"Receta encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}},"put":{"tags":["Gestión de Recetas"],"summary":"Actualizar receta","description":"Actualiza una receta existente","operationId":"updateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDTO"}}},"required":true},"responses":{"200":{"description":"Receta actualizada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"404":{"description":"Receta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}}},"/api/recipes":{"get":{"tags":["Gestión de Recetas"],"summary":"Listar recetas","description":"Obtiene todas las recetas activas","operationId":"getAllRecipes","responses":{"200":{"description":"Lista de recetas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recipe"}}}}}}},"post":{"tags":["Gestión de Recetas"],"summary":"Crear receta","description":"Crea una nueva receta de yogurt","operationId":"createRecipe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDTO"}}},"required":true},"responses":{"201":{"description":"Receta creada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}}},"/api/batches":{"get":{"tags":["yogurt-batch-controller"],"operationId":"getAllBatches","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PREPARING","HEATING","COOLING","INOCULATING","INCUBATING","REFRIGERATING","COMPLETED","FAILED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"post":{"tags":["yogurt-batch-controller"],"operationId":"startNewBatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartBatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/temperature":{"post":{"tags":["yogurt-batch-controller"],"operationId":"recordTemperature","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemperatureRecordDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/batches/{batchId}/refrigeration":{"post":{"tags":["yogurt-batch-controller"],"operationId":"startRefrigeration","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/inoculating":{"post":{"tags":["yogurt-batch-controller"],"operationId":"startInoculating","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/incubation":{"post":{"tags":["yogurt-batch-controller"],"operationId":"startIncubation","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/heating":{"post":{"tags":["yogurt-batch-controller"],"operationId":"startHeating","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/fail":{"post":{"tags":["yogurt-batch-controller"],"operationId":"markAsFailed","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/complete":{"post":{"tags":["yogurt-batch-controller"],"operationId":"completeBatch","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/recipes/{id}/deactivate":{"patch":{"tags":["Gestión de Recetas"],"summary":"Desactivar receta","operationId":"deactivateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/recipes/{id}/activate":{"patch":{"tags":["Gestión de Recetas"],"summary":"Activar receta","operationId":"activateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/recipes/search":{"get":{"tags":["Gestión de Recetas"],"summary":"Buscar recetas","description":"Busca recetas por palabra clave","operationId":"searchRecipes","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recipe"}}}}}}}},"/api/monitoring/dashboard":{"get":{"tags":["monitoring-controller"],"operationId":"getDashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}}}}},"/api/monitoring/batches/{batchId}/temperature":{"get":{"tags":["monitoring-controller"],"operationId":"getBatchTemperatureSummary","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemperatureSummary"}}}}}}},"/api/monitoring/batches/{batchId}/temperature-logs":{"get":{"tags":["monitoring-controller"],"operationId":"getTemperatureLogs","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TemperatureLog"}}}}}}}},"/api/monitoring/batches/active":{"get":{"tags":["monitoring-controller"],"operationId":"getActiveBatches","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}}},"/api/batches/{batchId}":{"get":{"tags":["yogurt-batch-controller"],"operationId":"getBatch","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}}},"components":{"schemas":{"IngredientDTO":{"type":"object","properties":{"name":{"type":"string"},"quantity":{"type":"number","format":"double"},"unit":{"type":"string"},"notes":{"type":"string"},"optional":{"type":"boolean"}},"description":"Lista de ingredientes de la receta"},"RecipeDTO":{"type":"object","properties":{"name":{"type":"string","description":"Nombre de la receta","example":"Yogurt natural"},"description":{"type":"string","description":"Descripción de la receta"},"defaultMilkVolume":{"type":"number","description":"Cantidad de leche en litros","format":"double","example":1.0},"defaultStarterAmount":{"type":"number","description":"Cantidad de cultivo inicial (starter)","format":"double","example":2.0},"heatingTemperature":{"type":"number","description":"Temperatura de calentamiento en °C","format":"double","example":85},"heatingDuration":{"type":"integer","description":"Duración del calentamiento en minutos","format":"int32","example":30},"inoculationTemperature":{"type":"number","description":"Temperatura de inoculación en °C","format":"double","example":45},"incubationTemperature":{"type":"number","description":"Temperatura de incubación en °C","format":"double","example":42},"minIncubationTime":{"type":"integer","description":"Tiempo mínimo de incubación en horas","format":"int32","example":6},"maxIncubationTime":{"type":"integer","description":"Tiempo máximo de incubación en horas","format":"int32","example":8},"refrigerationTime":{"type":"integer","description":"Tiempo de refrigeración en horas","format":"int32","example":4},"difficulty":{"type":"string","description":"Nivel de dificultad de la receta","example":"EASY","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"tips":{"type":"string","description":"Consejos adicionales para la preparación"},"ingredients":{"type":"array","description":"Lista de ingredientes de la receta","items":{"$ref":"#/components/schemas/IngredientDTO"}}},"description":"Objeto para crear o actualizar recetas de yogurt"},"Ingredient":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"quantity":{"type":"number","format":"double"},"unit":{"type":"string"},"notes":{"type":"string"},"optional":{"type":"boolean"}}},"Recipe":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/Ingredient"}},"defaultMilkVolume":{"type":"number","format":"double"},"defaultStarterAmount":{"type":"number","format":"double"},"heatingTemperature":{"type":"number","format":"double"},"heatingDuration":{"type":"integer","format":"int32"},"inoculationTemperature":{"type":"number","format":"double"},"incubationTemperature":{"type":"number","format":"double"},"minIncubationTime":{"type":"integer","format":"int32"},"maxIncubationTime":{"type":"integer","format":"int32"},"refrigerationTime":{"type":"integer","format":"int32"},"difficulty":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"tips":{"type":"string"},"active":{"type":"boolean"}}},"StartBatchRequest":{"type":"object","properties":{"recipeId":{"type":"integer","format":"int64"},"customMilkVolume":{"type":"number","format":"double"},"customStarterAmount":{"type":"number","format":"double"}}},"TemperatureLog":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"batch":{"$ref":"#/components/schemas/YogurtBatch"},"temperature":{"type":"number","format":"double"},"recordedAt":{"type":"string","format":"date-time"},"type":{"type":"string","enum":["HEATING","COOLING","INCUBATION","REFRIGERATION","MANUAL"]},"notes":{"type":"string"}}},"YogurtBatch":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"batchCode":{"type":"string"},"recipe":{"$ref":"#/components/schemas/Recipe"},"status":{"type":"string","enum":["PREPARING","HEATING","COOLING","INOCULATING","INCUBATING","REFRIGERATING","COMPLETED","FAILED"]},"milkVolume":{"type":"number","format":"double"},"starterAmount":{"type":"number","format":"double"},"targetTemperature":{"type":"number","format":"double"},"incubationTime":{"type":"integer","format":"int32"},"startTime":{"type":"string","format":"date-time"},"incubationStartTime":{"type":"string","format":"date-time"},"incubationEndTime":{"type":"string","format":"date-time"},"refrigerationStartTime":{"type":"string","format":"date-time"},"temperatureLogs":{"type":"array","items":{"$ref":"#/components/schemas/TemperatureLog"}},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"TemperatureRecordDTO":{"type":"object","properties":{"temperature":{"type":"number","format":"double"},"type":{"type":"string","enum":["HEATING","COOLING","INCUBATION","REFRIGERATION","MANUAL"]}}},"FailRequest":{"type":"object","properties":{"reason":{"type":"string"}}},"Dashboard":{"type":"object","properties":{"batchCounts":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"activeBatchesCount":{"type":"integer","format":"int64"},"completedToday":{"type":"integer","format":"int32"}}},"TemperatureSummary":{"type":"object","properties":{"currentTemperature":{"type":"number","format":"double"},"maximumTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"averageTemperature":{"type":"number","format":"double"}}}}}}